Move theme prefs to general prefs
